Move upload handler to separate file
